Add a comment explaining the use of double pointer
The push_head() function in linkedlist.c has a double pointer as an
argument. As a newbie, I couldn't understand it at first. The comment
contains a link to an explanation on StackOverflow. With this method you
get the new node without a return value, so it is easier and less prone
to bugs.